speech-recognition-python/requirements.txt
azure-cognitiveservices-speech
flask
